move getOverlappingTermsFromList into its own file
